backend/requirements.txt
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt>=5.3.0
Pillow==9.5.0
psycopg2-binary==2.9.6
django-filter
requests
cryptography
drf-spectacular
drf-spectacular-sidecar
google-auth
daphne>=4.0
channels>=4.0
channels-redis>=4.1
